separate components in post viewer control bar
